requirements.txt
flask
google-cloud-bigquery
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace